Move creatRow function to gmp/commands/dashboard
The createRow function has been duplicated. Now use a single place to
provide it.
